+++ RT1064/my-smartcar/CODE/motor.h
@@ -13,8 +13,8 @@
 #define PWM_4 PWM1_MODULE0_CHB_D13
 
 
-extern int16 encoder1,encoder2,encoder3,encoder4;
-extern int16 RC_encoder1,RC_encoder2,RC_encoder3,RC_encoder4;//滤波之后encoder的值
+extern int32 encoder1,encoder2,encoder3,encoder4;
+extern int32 RC_encoder1,RC_encoder2,RC_encoder3,RC_encoder4;//滤波之后encoder的值
 extern int32 duty1,duty2,duty3,duty4;
 extern int speed_tar_1,speed_tar_2,speed_tar_3,speed_tar_4;
 extern int speed_tar;
